# WinCCOAGraphQLRestApisExample/javascript/winccoa-graphql-server-main/restapi/openapi-full.yaml
openapi: 3.0.0
info:
  title: WinCC OA REST API
  version: 1.0.0
  description: |
    REST API for WinCC OA GraphQL Server - provides HTTP endpoints for all WinCC OA operations including:
    - Data point management (CRUD, value operations)
    - Data point type management
    - Tag queries with metadata
    - Alert handling
    - CNS (Central Navigation Service) operations
    - System information and redundancy
    - OPC UA configuration
  contact:
    name: API Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /restapi/v1/auth/login
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Missing required parameter
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
...
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Missing or invalid authentication token
    Forbidden:
      description: Insufficient permissions (admin role required)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
            message: Admin role required
paths:
...
  /restapi/v1/datapoints:
    get:
      tags: [Data Points]
      summary: List data points
      description: Search for data points by pattern
      security:
        - bearerAuth: []
      parameters:
        - name: pattern
          in: query
          schema:
            type: string
          description: Search pattern (wildcards supported)
          example: "ExampleDP_*"
        - name: dpType
          in: query
          schema:
            type: string
          description: Filter by data point type
        - name: ignoreCase
          in: query
          schema:
            type: boolean
          description: Case-insensitive search
      responses:
        '200':
          description: List of data points
          content:
            application/json:
              schema:
                type: object
                properties:
                  datapoints:
                    type: array
                    items:
                      type: string
                    example: ["ExampleDP_Arg1.", "ExampleDP_Arg2."]
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Data Points]
      summary: Create data point
      description: Create a new data point (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dpeName, dpType]
              properties:
                dpeName:
                  type: string
                  example: "MyDataPoint."
                dpType:
                  type: string
                  example: "ExampleDP_Arg"
                systemId:
                  type: integer
                  example: 1
                dpId:
                  type: integer
                  example: 100
      responses:
        '200':
          description: Data point created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # Data Points - Value Operations
  /restapi/v1/datapoints/{dpeName}/value:
    get:
      tags: [Data Points]
      summary: Get data point value
      description: Retrieve current value of a data point element
      security:
        - bearerAuth: []
      parameters:
        - name: dpeName
          in: path
          required: true
          schema:
            type: string
          description: Data point element name (URL encoded)
          example: "ExampleDP_Arg1.value"
      responses:
        '200':
          description: Current value
          content:
            application/json:
              schema:
                type: object
                properties:
                  value:
                    description: Current value (any type)
                    example: 42.5

    put:
      tags: [Data Points]
      summary: Set data point value
      description: Set the value of a data point element (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: dpeName
          in: path
          required: true
          schema:
            type: string
          description: Data point element name (URL encoded)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value:
                  description: New value (any type)
                  example: 50.0
      responses:
        '200':
          description: Value set successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...